fix folder structure
