--- react-chat/src/components/EditProfileHeader/index.ts
@@ -1,1 +1,0 @@
-export {default} from './EditProfileHeader';
