+++ react-chat/src/layouts/ChatLayout/ChatLayout.scss
