+++ react-chat/src/layouts/ChatLayout/ChatLayout.tsx
