+++ react-chat/src/layouts/HeaderLayout/HeaderLayout.scss
