+++ react-chat/src/modules/chat/ChatFooter/ChatFooter.tsx
