+++ react-chat/src/modules/chat/ChatFooter/index.ts
