+++ react-chat/src/modules/chat/ChatHeader/ChatHeader.scss
