+++ react-chat/src/modules/chat/ChatItem/ChatItem.scss
