+++ react-chat/src/modules/chat/ChatsHeader/ChatsHeader.scss
