+++ react-chat/src/modules/chat/ChatsList/ChatsList.scss
