+++ react-chat/src/modules/message/MessageItem/MessageItem.scss
