+++ react-chat/src/modules/message/MessagesList/MessagesList.scss
