+++ react-chat/src/modules/profile/ProfileHeader/ProfileHeader.scss
