+++ react-chat/src/modules/profile/ProfileHeader/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ProfileHeader';
